libnm: use macro and designated initializers for NMVariantAttributeSpec
I think initializing structs should (almost) be always done with designated
integers, because otherwise it's easy to get the order wrong. But what
is worse is that otherwise the order of fields gets additional meaning
not only about the memory layout, but also about the code that
initialize the structs.

Add a macro NM_VARIANT_ATTRIBUTE_SPEC_DEFINE() that replaces the other
(duplicate) macros. This macro also gets it right to mark the struct as
const.