device/trivial: add comment about lifetime of "kind" in tc_commit()
In general, all fields of public NMPlatform* structs must be
plain/simple. Meaning: copying the struct must return the same valid
value. That is why we intern the kind strings usually.

In general, if you have fields which lifetime is limited,
then the field cannot be inside the public part NMPlatform*.
See for example NMPClass.cmd_obj_copy() which can deep-copy an object.
But this is only suitable for private fields! The purpose of this
rule is that you always can safely copy a NMPlatform* struct.

This is violated in tc_commit() for the kind strings. The lifetime
of these strings is limited to the setting instance.

We also cannot intern the strings (because interned strings are leaked
indefinitely). We also cannot g_strdup() the strings, because the public
portions of NMPlatform* don't have a free/clone function that properly
frees/duplicates the string.

So, just add a comment how this works.