core: drop refcounting of NMAuthChain for simple flags
NMAuthChain is not really ref-counted, there is no need for that additional
complexity.

But it is graceful towards calling nm_auth_chain_destroy() from inside the
callback. The caller may do so.

But to track that we don't need a "ref_count". Two flags suffice: one to
say whether destroy was called and one to indicate that we are in the
process of finishing.

We already had the "done" flag that we used to indicate that the chain
is finished. So, we just need one more flag instead.